Move validation helpers into their own mixin
So we can reuse them on other forms.
