+++ components/ValidatingFormInput.vue
@@ -15,10 +15,10 @@
 </template>
 
 <script>
-import validationHelpers from '@/mixins/validationHelpers'
+import validationFieldHelpers from '@/mixins/validationFieldHelpers'
 
 export default {
-  mixins: [validationHelpers],
+  mixins: [validationFieldHelpers],
   computed: {
     maxLength() {
       return this.validationTypes.includes('maxLength')
